+++ src/ts/plugins/apiV3/risuai.d.ts
@@ -1877,6 +1877,7 @@ interface RisuaiPluginAPI {
     apiVersion: string
     platform: string
     saveMethod: string
+    deviceLocalPluginStorage: boolean
   }>
 
   /**